requirements.txt
python-dotenv==1.0.0
pandas==2.1.3
telethon==1.32.1
openai==1.3.5
google-generativeai==0.3.1
tiktoken==0.5.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
tqdm==4.66.1
requests==2.31.0
apscheduler==3.10.4
slack-sdk==3.26.0